+++ websweeper/src/components/GamePlay.js
@@ -1,0 +1,5 @@
+function playerClick() {
+  console.log('nothing3');
+}
+
+export default { playerClick };
