+++ websweeper/src/components/SetBoard.js
@@ -1,0 +1,10 @@
+
+function placeMines() {
+  console.log('nothing1');
+}
+
+function countMines() {
+    console.log('nothing2');
+}
+
+export { placeMines, countMines };
