+++ websweeper/src/css/web.css
@@ -6,7 +6,22 @@ body #board {
   padding-right: 100px;
 }
 
-/* surround bombs number color - determined at board generation */
+#homepage #startgame {
+  width: 300px; 
+  padding: 20px; 
+  cursor: pointer; 
+  box-shadow: 6px 6px 5px #999; 
+  -webkit-box-shadow: 6px 6px 5px #999; 
+  -moz-box-shadow: 6px 6px 5px #999; 
+  font-weight: bold; 
+  background: #ffff00; 
+  color: #000; 
+  border-radius: 10px; 
+  border: 1px solid #999; 
+  font-size: 150%;
+}
+
+/* surrounding bombs number color - determined at board generation */
 body #board .bomb-num1 {
   color: blue;
 }
