+++ websweeper/src/index.html
@@ -1,18 +1,18 @@
 <!DOCTYPE html>
 <html>
-    <title>Testing</title>
+    <title>Websweeper</title>
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="./css/css.css">
-    <title>Document</title>
+    <link rel="stylesheet" href="/css/web.css">
+    <title>Websweeper</title>
 </head>
 <body>
-  <div id="minesweeper">
     <div id="sizeSelection">
+      <script src="./components/WebSweeper.js" type="module"></script>
       <h1 class="sizeSelectionText">Select Your Board Size:</h1>
-      <form id="htmlform" onsubmit="getValue(); return false">
+      <form id="htmlform" onsubmit="return false;">
         <input id="small" type="radio" name="size" value="Small" checked>
         <span class="form-items">
           Small - (Best for beginner or a quick game)
@@ -30,8 +30,8 @@
       </form>
     </div>
     <div id="board">
-      <script src="./components/maketiles.js"></script>
+      <!-- Not sure if i need to include this script like this? Keeping here for now 
+      <script src="./components/WebSweeper.js"></script> -->
     </div>
-  </div>
 </body>
 </html>
